refactor: make parts A and B into nested packages
