+++ packages/riceclf/src/riceclf/Rice_Cammeo_Osmancik.arff
