+++ packages/riceclf/src/riceclf/__main__.py
@@ -1,0 +1,6 @@
+import sys
+
+from packages.riceclf.src.riceclf import main
+
+if __name__ == "__main__":
+    sys.exit(main())
