+++ packages/symreg/README.md
@@ -1,0 +1,1 @@
+# symreg
