+++ packages/symreg/src/symreg/ReadME.docx
