+++ packages/symreg/src/symreg/__main__.py
@@ -1,0 +1,6 @@
+import sys
+
+from packages.symreg.src.symreg import main
+
+if __name__ == "__main__":
+    sys.exit(main())
